feat(database): add fields to support prototype display requirements
Database changes:
- osg_student: add target_position, last_login_time
- osg_contract: add used_hours, remaining_hours

Domain model updates:
- OsgStudent.java: add targetPosition (String), lastLoginTime (Date)
- OsgContract.java: change usedHours/remainingHours from BigDecimal to Integer

Mapper XML updates:
- OsgStudentMapper.xml: include new fields in resultMap and selectStudentColumns
- OsgContractMapper.xml: use actual used_hours/remaining_hours columns instead of computed values

Service layer updates:
- OsgContractServiceImpl.java: update type handling for Integer hours
- OsgStudentServiceImpl.java: update type handling for Integer hours

Ticket: T-176
Story: S-038

+++ ruoyi-system/src/main/java/com/ruoyi/system/domain/OsgContract.java
@@ -28,9 +28,9 @@ public class OsgContract extends BaseEntity
 
     private Integer totalHours;
 
-    private BigDecimal usedHours;
+    private Integer usedHours;
 
-    private BigDecimal remainingHours;
+    private Integer remainingHours;
 
     private Date startDate;
 
@@ -132,22 +132,22 @@ public class OsgContract extends BaseEntity
         this.totalHours = totalHours;
     }
 
-    public BigDecimal getUsedHours()
+    public Integer getUsedHours()
     {
         return usedHours;
     }
 
-    public void setUsedHours(BigDecimal usedHours)
+    public void setUsedHours(Integer usedHours)
     {
         this.usedHours = usedHours;
     }
 
-    public BigDecimal getRemainingHours()
+    public Integer getRemainingHours()
     {
         return remainingHours;
     }
 
-    public void setRemainingHours(BigDecimal remainingHours)
+    public void setRemainingHours(Integer remainingHours)
     {
         this.remainingHours = remainingHours;
     }
